Улучшает структуру кода, убирает лишнию информацию в координатах
