+++ js/main.js
@@ -1,6 +1,6 @@
 import './fill-data/fill-template.js';
 import './toggle-state-page.js';
-import './validation.js';
+import './validation/validation.js';
 import './map/map.js';
 import './slider.js';
