+++ js/validation/form.js
@@ -1,0 +1,1 @@
+export const form = document.querySelector('.ad-form');
